update - attempt to implement combineReducers, debug needed
