+++ smurfs/src/components/App.js
@@ -18,6 +18,7 @@ const App = props => {
   useEffect(() => {
     let initSmurfs = props.getSmurfs();
     setNewSmurfs(initSmurfs);
+    console.log(newSmurfs);
   }, []);
 
   console.log("This is props at App.js: ", props);
