+++ smurfs/src/index.js
@@ -9,7 +9,7 @@ import { composeWithDevTools } from "redux-devtools-extension";
 import { createStore, applyMiddleware } from "redux";
 import thunk from "redux-thunk";
 import { Provider } from "react-redux";
-import { rootReducer } from "./Redux/reducers/index";
+import rootReducer from "./Redux/reducers/index";
 
 import "./App.css";
